+++ tests/BusinessLogic/Common/TestComponents/Order/TestOrderRepository.php
@@ -318,22 +318,6 @@ class TestOrderRepository implements OrderRepository
     }
 
     /**
-     * Returns whether shipment identified by provided reference has Packlink shipment labels set.
-     *
-     * @param string $shipmentReference Packlink shipment reference.
-     *
-     * @return bool Returns TRUE if labels are set; otherwise returns FALSE.
-     */
-    public function isLabelSet($shipmentReference)
-    {
-        $order = $this->getOrder($shipmentReference, 0, '', false);
-
-        $packlinkShipmentLabels = $order->getPacklinkShipmentLabels();
-
-        return !empty($packlinkShipmentLabels);
-    }
-
-    /**
      * Retrieves list of order references where order is in one of the provided statuses.
      *
      * @param array $statuses List of order statuses.